+++ tests/test_delta.py
@@ -11,7 +11,7 @@ from deepdiff.helper import np, number_to_string, TEXT_VIEW, DELTA_VIEW, CannotC
 from deepdiff.path import GETATTR, GET
 from deepdiff.delta import (
     ELEM_NOT_FOUND_TO_ADD_MSG,
-    VERIFICATION_MSG, VERIFY_SYMMETRY_MSG, not_found, DeltaNumpyOperatorOverrideError,
+    VERIFICATION_MSG, VERIFY_BIDIRECTIONAL_MSG, not_found, DeltaNumpyOperatorOverrideError,
     BINIARY_MODE_NEEDED_MSG, DELTA_AT_LEAST_ONE_ARG_NEEDED, DeltaError,
     INVALID_ACTION_WHEN_CALLING_GET_ELEM, INVALID_ACTION_WHEN_CALLING_SIMPLE_SET_ELEM,
     INVALID_ACTION_WHEN_CALLING_SIMPLE_DELETE_ELEM, INDEXES_NOT_FOUND_WHEN_IGNORE_ORDER,
@@ -35,6 +35,11 @@ class TestBasicsOfDelta:
         delta2 = Delta(dump, deserializer=json_loads)
         assert delta2 + t1 == t2
         assert t1 + delta2 == t2
+        with pytest.raises(ValueError) as exc_info:
+            t2 - delta
+        assert 'Please recreate the delta with bidirectional=True' == str(exc_info.value)
+        delta = Delta(diff, serializer=json_dumps, bidirectional=True)
+        assert t2 - delta == t1
 
     def test_to_null_delta1_json(self):
         t1 = 1
@@ -71,6 +76,8 @@ class TestBasicsOfDelta:
         ]
 
         assert flat_expected1 == flat_result1
+        delta2 = Delta(diff=diff, bidirectional=True)
+        assert t1 == t2 - delta2
 
     def test_list_difference_dump_delta(self):
         t1 = [1, 2]
@@ -245,13 +252,13 @@ class TestBasicsOfDelta:
         # since we sort the keys by the path elements, root[3] is gonna be processed before root[20]
         expected_msg = ELEM_NOT_FOUND_TO_ADD_MSG.format(3, 'root[3]')
 
-        delta2 = Delta(diff, verify_symmetry=True, raise_errors=True, log_errors=False)
+        delta2 = Delta(diff, bidirectional=True, raise_errors=True, log_errors=False)
         with pytest.raises(ValueError) as excinfo:
             delta2 + t1
         assert expected_msg == str(excinfo.value)
         assert not mock_logger.called
 
-        delta3 = Delta(diff, verify_symmetry=True, raise_errors=True, log_errors=True)
+        delta3 = Delta(diff, bidirectional=True, raise_errors=True, log_errors=True)
         with pytest.raises(ValueError) as excinfo:
             delta3 + t1
         assert expected_msg == str(excinfo.value)
@@ -289,6 +296,9 @@ class TestBasicsOfDelta:
 
         assert flat_expected1 == flat_result1
 
+        delta2 = Delta(diff=diff, bidirectional=True)
+        assert t1 == t2 - delta2
+
     def test_list_difference_delta_raises_error_if_prev_value_does_not_match(self):
         t1 = [1, 2, 6]
         t2 = [1, 3, 2, 5]
@@ -308,14 +318,14 @@ class TestBasicsOfDelta:
             }
         }
 
-        expected_msg = VERIFICATION_MSG.format('root[2]', 5, 6, VERIFY_SYMMETRY_MSG)
+        expected_msg = VERIFICATION_MSG.format('root[2]', 5, 6, VERIFY_BIDIRECTIONAL_MSG)
 
-        delta = Delta(diff, verify_symmetry=True, raise_errors=True)
+        delta = Delta(diff, bidirectional=True, raise_errors=True)
         with pytest.raises(ValueError) as excinfo:
             delta + t1
         assert expected_msg == str(excinfo.value)
 
-        delta2 = Delta(diff, verify_symmetry=False)
+        delta2 = Delta(diff, bidirectional=False)
         assert delta2 + t1 == t2
 
         flat_result2 = delta2.to_flat_dicts()
@@ -357,6 +367,9 @@ class TestBasicsOfDelta:
 
         assert flat_expected == flat_result
 
+        delta2 = Delta(diff=diff, bidirectional=True)
+        assert t1 == t2 - delta2
+
     @mock.patch('deepdiff.delta.logger.error')
     def test_list_difference_delta_if_item_is_already_removed(self, mock_logger):
         t1 = [1, 2, 'to_be_removed']
@@ -369,12 +382,12 @@ class TestBasicsOfDelta:
             }
         }
         expected_msg = VERIFICATION_MSG.format("root[3]", 'to_be_removed2', not_found, 'list index out of range')
-        delta = Delta(diff, verify_symmetry=True, raise_errors=True)
+        delta = Delta(diff, bidirectional=True, raise_errors=True)
         with pytest.raises(DeltaError) as excinfo:
             delta + t1
         assert expected_msg == str(excinfo.value)
 
-        delta2 = Delta(diff, verify_symmetry=False, raise_errors=False)
+        delta2 = Delta(diff, bidirectional=False, raise_errors=False)
         assert t1 + delta2 == t2
         expected_msg = UNABLE_TO_GET_PATH_MSG.format('root[3]')
         mock_logger.assert_called_with(expected_msg)
@@ -397,14 +410,14 @@ class TestBasicsOfDelta:
                 "root[4]['b'][3]": 'to_be_removed2'
             }
         }
-        expected_msg = VERIFICATION_MSG.format("root[4]['b'][2]", 'to_be_removed', 'wrong', VERIFY_SYMMETRY_MSG)
+        expected_msg = VERIFICATION_MSG.format("root[4]['b'][2]", 'to_be_removed', 'wrong', VERIFY_BIDIRECTIONAL_MSG)
 
-        delta = Delta(diff, verify_symmetry=True, raise_errors=True)
+        delta = Delta(diff, bidirectional=True, raise_errors=True)
         with pytest.raises(ValueError) as excinfo:
             delta + t1
         assert expected_msg == str(excinfo.value)
 
-        delta2 = Delta(diff, verify_symmetry=False, raise_errors=True)
+        delta2 = Delta(diff, bidirectional=False, raise_errors=True)
         assert t1 + delta2 == t2
 
     def test_delta_dict_items_added_retain_order(self):
@@ -434,13 +447,16 @@ class TestBasicsOfDelta:
         diff = DeepDiff(t1, t2)
         delta_dict = diff._to_delta_dict()
         assert expected_delta_dict == delta_dict
-        delta = Delta(diff, verify_symmetry=False, raise_errors=True)
+        delta = Delta(diff, bidirectional=False, raise_errors=True)
 
         result = t1 + delta
         assert result == t2
 
         assert list(result.keys()) == [6, 7, 3, 5, 2, 4]
         assert list(result.keys()) == list(t2.keys())
+
+        delta2 = Delta(diff=diff, bidirectional=True)
+        assert t1 == t2 - delta2
 
 
 picklalbe_obj_without_item = PicklableClass(11)
@@ -701,8 +717,11 @@ class TestDelta:
         diff = DeepDiff(t1, t2, **deepdiff_kwargs)
         delta_dict = diff._to_delta_dict(**to_delta_kwargs)
         assert expected_delta_dict == delta_dict, f"test_delta_cases {test_name} failed."
-        delta = Delta(diff, verify_symmetry=False, raise_errors=True)
+        delta = Delta(diff, bidirectional=False, raise_errors=True)
         assert t1 + delta == t2, f"test_delta_cases {test_name} failed."
+
+        delta2 = Delta(diff, bidirectional=True, raise_errors=True)
+        assert t2 - delta2 == t1, f"test_delta_cases {test_name} failed."
 
 
 DELTA_IGNORE_ORDER_CASES = {
@@ -997,7 +1016,7 @@ class TestIgnoreOrderDelta:
         diff = DeepDiff(t1, t2, **deepdiff_kwargs)
         delta_dict = diff._to_delta_dict(**to_delta_kwargs)
         assert expected_delta_dict == delta_dict, f"test_ignore_order_delta_cases {test_name} failed"
-        delta = Delta(diff, verify_symmetry=False, raise_errors=True)
+        delta = Delta(diff, bidirectional=False, raise_errors=True)
         expected_t1_plus_delta = t2 if expected_t1_plus_delta == 't2' else expected_t1_plus_delta
         t1_plus_delta = t1 + delta
         assert t1_plus_delta == expected_t1_plus_delta, f"test_ignore_order_delta_cases {test_name} failed: diff = {DeepDiff(t1_plus_delta, expected_t1_plus_delta, ignore_order=True)}"
@@ -1159,7 +1178,7 @@ class TestNumpyDelta:
         delta_dict = diff._to_delta_dict(**to_delta_kwargs)
         if expected_delta_dict:
             assert expected_delta_dict == delta_dict, f"test_numpy_delta_cases {test_name} failed."
-        delta = Delta(diff, verify_symmetry=False, raise_errors=True)
+        delta = Delta(diff, bidirectional=False, raise_errors=True)
         if expected_result == 't2':
             result = delta + t1
             assert np.array_equal(result, t2), f"test_numpy_delta_cases {test_name} failed."
@@ -1300,7 +1319,7 @@ class TestDeltaOther:
         Test a specific case where path was a list of elements (in the form of tuples)
         and the item could not be found.
         """
-        delta = Delta({}, verify_symmetry=True, raise_errors=True, log_errors=False)
+        delta = Delta({}, bidirectional=True, raise_errors=True, log_errors=False)
         with pytest.raises(DeltaError) as excinfo:
             delta._get_elem_and_compare_to_old_value(
                 obj={},
@@ -1378,7 +1397,7 @@ class TestDeltaOther:
         delta2_again = Delta(flat_dict_list=flat_expected2)
         assert delta2.diff == delta2_again.diff
 
-        delta3 = Delta(diff, raise_errors=False, verify_symmetry=True)
+        delta3 = Delta(diff, raise_errors=False, bidirectional=True)
         flat_result3 = delta3.to_flat_dicts()
         flat_expected3 = [{'path': [1, 2, 0], 'action': 'values_changed', 'value': 5, 'old_value': 4}]
         assert flat_expected3 == flat_result3
@@ -1405,11 +1424,11 @@ class TestDeltaOther:
 
         diff = DeepDiff(t1, t2)
 
-        delta2 = Delta(diff, raise_errors=False, verify_symmetry=True)
+        delta2 = Delta(diff, raise_errors=False, bidirectional=True)
         t4 = delta2 + t3
 
         assert [2] == t4
-        expected_msg = VERIFICATION_MSG.format('root[0]', 1, 3, VERIFY_SYMMETRY_MSG)
+        expected_msg = VERIFICATION_MSG.format('root[0]', 1, 3, VERIFY_BIDIRECTIONAL_MSG)
         mock_logger.assert_called_once_with(expected_msg)
 
     @mock.patch('deepdiff.delta.logger.error')
@@ -1420,7 +1439,7 @@ class TestDeltaOther:
 
         diff = DeepDiff(t1, t2, ignore_order=True, report_repetition=True)
 
-        delta2 = Delta(diff, raise_errors=False, verify_symmetry=True)
+        delta2 = Delta(diff, raise_errors=False, bidirectional=True)
         t4 = delta2 + t3
 
         assert [5] == t4
@@ -1435,7 +1454,7 @@ class TestDeltaOther:
 
         diff = DeepDiff(t1, t2, ignore_order=True, report_repetition=True)
 
-        delta = Delta(diff, raise_errors=False, verify_symmetry=True)
+        delta = Delta(diff, raise_errors=False, bidirectional=True)
         t4 = delta + t3
 
         assert [1, 2, 'C'] == t4
@@ -1450,13 +1469,13 @@ class TestDeltaOther:
 
         diff = DeepDiff(t1, t2, ignore_order=True, report_repetition=True)
 
-        # when verify_symmetry=False, we still won't remove the item that is different
+        # when bidirectional=False, we still won't remove the item that is different
         # than what we expect specifically when ignore_order=True when generating the diff.
         # The reason is that when ignore_order=True, we can' rely too much on the index
         # of the item alone to delete it. We need to make sure we are deleting the correct value.
-        # The expected behavior is exactly the same as when verify_symmetry=True
+        # The expected behavior is exactly the same as when bidirectional=True
         # specifically for when ignore_order=True AND an item is removed.
-        delta = Delta(diff, raise_errors=False, verify_symmetry=False)
+        delta = Delta(diff, raise_errors=False, bidirectional=False)
         t4 = delta + t3
 
         assert [1, 2, 'C'] == t4
@@ -1470,7 +1489,7 @@ class TestDeltaOther:
         t3 = {}
 
         diff = DeepDiff(t1, t2, ignore_order=True, report_repetition=True)
-        delta = Delta(diff, raise_errors=False, verify_symmetry=False)
+        delta = Delta(diff, raise_errors=False, bidirectional=False)
         t4 = delta + t3
 
         assert {} == t4
@@ -1481,7 +1500,7 @@ class TestDeltaOther:
         t1 = [1, 2, 'B']
         t2 = [1, 2]
         diff = DeepDiff(t1, t2, ignore_order=True, report_repetition=True)
-        delta = Delta(diff, raise_errors=False, verify_symmetry=False)
+        delta = Delta(diff, raise_errors=False, bidirectional=False)
 
         result = delta.to_dict()
         expected = {'iterable_items_removed_at_indexes': {'root': {2: 'B'}}}
@@ -1549,7 +1568,11 @@ class TestDeltaOther:
         delta_again = Delta(flat_dict_list=flat_expected)
         assert delta.diff == delta_again.diff
 
-        flat_result2 = delta.to_flat_dicts(report_type_changes=False)
+        with pytest.raises(ValueError) as exc_info:
+            delta.to_flat_dicts(report_type_changes=False)
+        assert str(exc_info.value).startswith("When converting to flat dictionaries, if report_type_changes=False and there are type")
+        delta2 = Delta(dump, always_include_values=True)
+        flat_result2 = delta2.to_flat_dicts(report_type_changes=False)
         flat_expected2 = [{'path': ['a'], 'action': 'values_changed', 'value': 1}]
         assert flat_expected2 == flat_result2
 
@@ -1876,7 +1899,7 @@ class TestDeltaCompareFunc:
         t1 = {"field1": {"joe": "Joe"}}
         t2 = {"field1": {"joe": "Joe Nobody"}, "field2": {"jimmy": "Jimmy"}}
         diff = DeepDiff(t1, t2)
-        delta = Delta(diff=diff)
+        delta = Delta(diff=diff, always_include_values=True)
         flat_result = delta.to_flat_dicts(report_type_changes=False)
         flat_expected = [
             {'path': ['field2', 'jimmy'], 'value': 'Jimmy', 'action': 'dictionary_item_added'},
@@ -1896,7 +1919,7 @@ class TestDeltaCompareFunc:
         t1 = {"field1": {"joe": "Joe"}}
         t2 = {"field1": {"joe": "Joe Nobody"}, "field2": {"jimmy": "Jimmy", "sar": "Sarah"}}
         diff = DeepDiff(t1, t2)
-        delta = Delta(diff=diff)
+        delta = Delta(diff=diff, always_include_values=True)
         flat_result = delta.to_flat_dicts(report_type_changes=False)
         flat_expected = [
             {'path': ['field2'], 'value': {'jimmy': 'Jimmy', 'sar': 'Sarah'}, 'action': 'dictionary_item_added'},
@@ -1912,7 +1935,7 @@ class TestDeltaCompareFunc:
         t2 = {"field1": {"joe": "Joe"}, "field2": ["James"]}
         t3 = {"field1": {"joe": "Joe"}, "field2": ["James", "Jack"]}
         diff = DeepDiff(t1, t2)
-        delta = Delta(diff=diff)
+        delta = Delta(diff=diff, always_include_values=True)
         flat_result = delta.to_flat_dicts(report_type_changes=False)
         flat_expected = [{'path': ['field2', 0], 'value': 'James', 'action': 'iterable_item_added'}]
         assert flat_expected == flat_result
@@ -1922,7 +1945,7 @@ class TestDeltaCompareFunc:
         assert t2 == t1 + delta_again
 
         diff2 = DeepDiff(t2, t3)
-        delta2 = Delta(diff=diff2)
+        delta2 = Delta(diff=diff2, always_include_values=True)
         flat_result2 = delta2.to_flat_dicts(report_type_changes=False)
         flat_expected2 = [{'path': ['field2', 1], 'value': 'Jack', 'action': 'iterable_item_added'}]
         assert flat_expected2 == flat_result2
@@ -1937,7 +1960,7 @@ class TestDeltaCompareFunc:
         t2 = {"field1": {"joe": "Joe"}, "field2": {"James"}}
         t3 = {"field1": {"joe": "Joe"}, "field2": {"James", "Jack"}}
         diff = DeepDiff(t1, t2)
-        delta = Delta(diff=diff)
+        delta = Delta(diff=diff, always_include_values=True)
         assert t2 == t1 + delta
         flat_result = delta.to_flat_dicts(report_type_changes=False)
         flat_expected = [{'path': ['field2'], 'value': 'James', 'action': 'set_item_added'}]
@@ -1948,7 +1971,7 @@ class TestDeltaCompareFunc:
         assert t2 == t1 + delta_again
 
         diff = DeepDiff(t2, t3)
-        delta2 = Delta(diff=diff)
+        delta2 = Delta(diff=diff, always_include_values=True)
         flat_result2 = delta2.to_flat_dicts(report_type_changes=False)
         flat_expected2 = [{'path': ['field2'], 'value': 'Jack', 'action': 'set_item_added'}]
         assert flat_expected2 == flat_result2
@@ -1962,7 +1985,7 @@ class TestDeltaCompareFunc:
         t2 = {"field1": {"joe": "Joe"}, "field2": ("James", )}
         t3 = {"field1": {"joe": "Joe"}, "field2": ("James", "Jack")}
         diff = DeepDiff(t1, t2)
-        delta = Delta(diff=diff)
+        delta = Delta(diff=diff, always_include_values=True)
         flat_expected = delta.to_flat_dicts(report_type_changes=False)
         expected_result = [{'path': ['field2', 0], 'value': 'James', 'action': 'iterable_item_added'}]
         assert expected_result == flat_expected
@@ -1971,7 +1994,7 @@ class TestDeltaCompareFunc:
         assert {'iterable_items_added_at_indexes': {"root['field2']": {0: 'James'}}} == delta_again.diff
 
         diff = DeepDiff(t2, t3)
-        delta2 = Delta(diff=diff)
+        delta2 = Delta(diff=diff, always_include_values=True)
         flat_result2 = delta2.to_flat_dicts(report_type_changes=False)
         expected_result2 = [{'path': ['field2', 1], 'value': 'Jack', 'action': 'iterable_item_added'}]
         assert expected_result2 == flat_result2
@@ -1983,12 +2006,12 @@ class TestDeltaCompareFunc:
         t1 = {"field1": {"joe": "Joe"}}
         t2 = {"field1": {"joe": "Joe"}, "field2": ["James", "Jack"]}
         diff = DeepDiff(t1, t2)
-        delta = Delta(diff=diff)
+        delta = Delta(diff=diff, always_include_values=True)
         flat_result = delta.to_flat_dicts(report_type_changes=False)
         expected_result = [{'path': ['field2'], 'value': ['James', 'Jack'], 'action': 'dictionary_item_added'}]
         assert expected_result == flat_result
 
-        delta2 = Delta(diff=diff, verify_symmetry=True)
+        delta2 = Delta(diff=diff, bidirectional=True, always_include_values=True)
         flat_result2 = delta2.to_flat_dicts(report_type_changes=False)
         assert expected_result == flat_result2
 
@@ -1999,10 +2022,38 @@ class TestDeltaCompareFunc:
         t1 = picklalbe_obj_without_item
         t2 = PicklableClass(10)
         diff = DeepDiff(t1, t2)
-        delta = Delta(diff=diff)
+        delta = Delta(diff=diff, always_include_values=True)
         flat_result = delta.to_flat_dicts(report_type_changes=False)
         expected_result = [{'path': ['item'], 'value': 10, 'action': 'attribute_added'}]
         assert expected_result == flat_result
 
         delta_again = Delta(flat_dict_list=flat_result)
         assert delta.diff == delta_again.diff
+
+    def test_flatten_when_simple_type_change(self):
+        t1 = [1, 2, '3']
+        t2 = [1, 2, 3]
+
+        diff = DeepDiff(t1, t2)
+        expected_diff = {
+            'type_changes': {'root[2]': {'old_type': str, 'new_type': int, 'old_value': '3', 'new_value': 3}}
+        }
+
+        assert expected_diff == diff
+        delta = Delta(diff=diff)
+        with pytest.raises(ValueError) as exc_info:
+            delta.to_flat_dicts(report_type_changes=False)
+        assert str(exc_info.value).startswith("When converting to flat dictionaries")
+
+        delta2 = Delta(diff=diff, always_include_values=True)
+        flat_result2 = delta2.to_flat_dicts(report_type_changes=False)
+        expected_result2 = [{'path': [2], 'action': 'values_changed', 'value': 3}]
+        assert expected_result2 == flat_result2
+
+        delta3 = Delta(diff=diff, always_include_values=True, bidirectional=True)
+        flat_result3 = delta3.to_flat_dicts(report_type_changes=False)
+        expected_result3 = [{'path': [2], 'action': 'values_changed', 'value': 3, 'old_value': '3'}]
+        assert expected_result3 == flat_result3
+
+        delta_again = Delta(flat_dict_list=flat_result3)
+        assert {'values_changed': {'root[2]': {'new_value': 3, 'old_value': '3'}}} == delta_again.diff